Make cells public simple array for saftey & simplicity
No invarients need to be maintained, so it's
safe to make Matrix POD.
Initialization is now simpler and safer, since proper
aggregte init can be used. By way of example, the
change caught an instance of to many items in an
initializer list in one test case.

Keeping subscripting
operator() for convenience.