Implement rendering of shadows
Unfortunately, this uncovers a problem with using float instead of
double, as acne is really bad unless the over_point value is offset by
1000 x epsilon, rather than just epsilon. A bit of research indicates
this is probably a precision issue, and the best bet is probably to
move to doubles instead of floats throughout. Should have
templated that type, I guess. :-(